refactor : provider 개념 독립
