+++ product/src/main/java/com/emotionalcart/product/presentation/dto/ReadProviders.java
@@ -1,8 +1,6 @@
 package com.emotionalcart.product.presentation.dto;
 
-import java.time.LocalDateTime;
-
-import com.emotionalcart.core.feature.product.Provider;
+import com.emotionalcart.core.feature.provider.Provider;
 
 import lombok.Data;
 
@@ -10,12 +8,12 @@ public class ReadProviders {
 
     @Data
     public static class Response {
-        private Long providerId;
+        private Long id;
         private String name;
         private String description;
 
         public Response(Provider provider) {
-            this.providerId = provider.getId();
+            this.id = provider.getId();
             this.name = provider.getName();
             this.description = provider.getDescription();
         }
